Modify HasSnakeHitObstacle

+++ include/border.h
@@ -74,7 +74,7 @@ namespace snake {
   private:
    // Determine how precise do we want the game to be
    // In other words, how close do we need the snake to be with the treat in order to eat it
-   const size_t kGameFlexibilityConstant = 8;
+   const float kGameFlexibilityConstant = 8.0f;
 
    // Create container
    const char* kContainerColor = "white";
@@ -97,7 +97,7 @@ namespace snake {
 
    // Create Obstacles
    std::vector<Obstacle> obstacles_;
-   const size_t KObstacleSideLength = 40;
+   const float KObstacleSideLength = 40.0f;
    const size_t kNumberOfObstacles = 3;
 
    // Game over text
